feat: Move subcomponent view models in the subcomponents folder
