+++ GovUkDesignSystem.SnapshotTests/GovUkDesignSystemComponents/CookieBannerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using GovUkDesignSystem.GovUkDesignSystemComponents;
+using GovUkDesignSystem.GovUkDesignSystemComponents.SubComponents;
 using GovUkDesignSystem.SnapshotTests.Helpers;
 using Xunit;
