+++ GovUkDesignSystem/GovUkDesignSystemComponents/SubComponents/CookieBannerMessage.cshtml
@@ -1,6 +1,6 @@
 ﻿@using GovUkDesignSystem.GovUkDesignSystemComponents.SubComponents
 @using GovUkDesignSystem.Helpers
-@model GovUkDesignSystem.GovUkDesignSystemComponents.CookieBannerMessageViewModel
+@model GovUkDesignSystem.GovUkDesignSystemComponents.SubComponents.CookieBannerMessageViewModel
 
 <div class="govuk-cookie-banner__message govuk-width-container @Model.Classes"
      @(Model.Role is not null ? $"role=\"{Model.Role}\"" : "")
