+++ GovUkDesignSystem/GovUkDesignSystemComponents/SubComponents/CookieBannerMessageAction.cshtml
@@ -1,7 +1,7 @@
 ﻿@using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
 @using GovUkDesignSystem.Helpers
-@model GovUkDesignSystem.GovUkDesignSystemComponents.CookieBannerMessageActionViewModel
+@model GovUkDesignSystem.GovUkDesignSystemComponents.SubComponents.CookieBannerMessageActionViewModel
 
 @if (Model.Href is not null)
 {
